Fix for JEXL-131; modified the code as mentioned in the mailing list thread. (http://apache-commons.680414.n4.nabble.com/jexl-Unified-jexl-throws-exception-when-processing-white-space-in-file-tt4578778.html); Fix for method exception not being reported correctly, backport of jexl3 behavior; Fix for variable/property existence versus null operands raising the correct exceptions: added separate strict flag to JexlEngine so JexlArithmetic can have its own , backport of jexl3 behavior. Fix for decimal number constants not defaulting to "double" behavior ( aka Grace's JEXL-126 fixed in jexl3 backport); Modified pom for 2.1.2-SNAPSHOT;
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/jexl/branches/2.0@1332806 13f79535-47bb-0310-9956-ffa450edef68

+++ src/test/java/org/apache/commons/jexl2/ArithmeticTest.java
@@ -122,7 +122,7 @@ public class ArithmeticTest extends JexlTestCase {
         asserter.setVariable("aBigDecimal", new BigDecimal("8.8"));
 
         asserter.assertExpression("-3", new Integer("-3"));
-        asserter.assertExpression("-3.0", new Float("-3.0"));
+        asserter.assertExpression("-3.0", new Double("-3.0"));
         asserter.assertExpression("-aByte", new Byte((byte) -1));
         asserter.assertExpression("-aShort", new Short((short) -2));
         asserter.assertExpression("-anInteger", new Integer(-3));
@@ -258,7 +258,7 @@ public class ArithmeticTest extends JexlTestCase {
         MatchingContainer ad = new MatchingContainer(ai);
         Set<Integer> as = ad.values;
         Object[] vars = { ai, al, am, ad, as };
-      
+
         for(Object var : vars) {
             asserter.setVariable("container", var);
             for(int x : ai) {
